refactor: expose standalone indexer runner
