+++ indexer-standalone/src/lib.rs
@@ -1,0 +1,217 @@
+// This file is part of midnight-indexer.
+// Copyright (C) Midnight Foundation
+// SPDX-License-Identifier: Apache-2.0
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#[cfg(feature = "standalone")]
+mod config;
+
+#[cfg(feature = "standalone")]
+use config::{Config, InfraConfig};
+
+/// Run the standalone indexer using its conventional configuration sources.
+#[cfg(feature = "standalone")]
+pub fn run() -> anyhow::Result<()> {
+    use anyhow::Context;
+    use indexer_common::config::ConfigExt;
+
+    let config = Config::load().context("load configuration")?;
+    run_with_config(config, true)
+}
+
+/// Run the indexer inside another process using an explicit configuration file.
+///
+/// The host owns process-wide logging and tracing. Indexer metrics retain their existing
+/// configuration because they use a separate exporter and listener.
+#[cfg(feature = "standalone")]
+pub fn run_embedded(config_path: impl AsRef<std::path::Path>) -> anyhow::Result<()> {
+    use anyhow::Context;
+    use indexer_common::config::ConfigExt;
+
+    let config = Config::load_from(config_path).context("load embedded indexer configuration")?;
+    run_with_config(config, false)
+}
+
+#[cfg(feature = "standalone")]
+fn run_with_config(config: Config, initialize_tracing: bool) -> anyhow::Result<()> {
+    use anyhow::Context;
+    use chain_indexer::{
+        application as chain_app,
+        infra::{storage as chain_storage, subxt_node::SubxtNode},
+    };
+    use indexer_api::{
+        application as api_app,
+        infra::{api::AxumApi, storage as api_storage},
+    };
+    use indexer_common::{
+        cipher::make_cipher,
+        infra::{ledger_db, migrations, pool, pub_sub},
+        telemetry,
+    };
+    use log::{info, warn};
+    use spo_indexer::{
+        application as spo_app,
+        infra::{spo_client::SPOClient, storage as spo_storage},
+    };
+    use tokio::{
+        runtime::Builder,
+        select,
+        signal::unix::{SignalKind, signal},
+        task,
+    };
+    use wallet_indexer::{application as wallet_app, infra::storage as wallet_storage};
+
+    let Config {
+        thread_stack_size,
+        application_config,
+        spo_config,
+        infra_config,
+        telemetry_config:
+            telemetry::Config {
+                tracing_config,
+                metrics_config,
+            },
+    } = config;
+
+    info!(
+        application_config:?,
+        infra_config:?;
+        "starting"
+    );
+
+    let ledger_state_retention = application_config.ledger_state_retention.get();
+    let max_snapshot_age = infra_config
+        .api_config
+        .subscription_config
+        .dust_generations
+        .max_snapshot_age;
+    assert!(
+        u64::try_from(ledger_state_retention).unwrap_or(u64::MAX) > u64::from(max_snapshot_age),
+        "ledger_state_retention ({ledger_state_retention}) must exceed \
+         dust_generations.max_snapshot_age ({max_snapshot_age}): otherwise a snapshot can pass \
+         the freshness check yet resolve to garbage-collected ledger state and panic on load"
+    );
+
+    let InfraConfig {
+        run_migrations,
+        storage_config,
+        ledger_db_config,
+        node_config,
+        spo_node_config,
+        api_config,
+        secret,
+    } = infra_config;
+
+    let runtime = Builder::new_multi_thread()
+        .enable_all()
+        .thread_stack_size(thread_stack_size as usize)
+        .build()
+        .context("build Tokio runtime")?;
+
+    runtime.block_on(async {
+        if initialize_tracing {
+            telemetry::init_tracing(tracing_config);
+        } else if tracing_config.enabled {
+            warn!("embedded indexer tracing is disabled because the host owns global tracing");
+        }
+        telemetry::init_metrics(metrics_config);
+
+        let pool = pool::sqlite::SqlitePool::new(storage_config)
+            .await
+            .context("create DB pool for Sqlite")?;
+        if run_migrations {
+            migrations::sqlite::run(&pool)
+                .await
+                .context("run Sqlite migrations")?;
+        }
+
+        let cipher = make_cipher(secret).context("make cipher")?;
+        let pub_sub = pub_sub::in_mem::InMemPubSub::default();
+
+        ledger_db::init(ledger_db_config)
+            .await
+            .context("initialize ledger db")?;
+
+        let chain_indexer = {
+            let storage = chain_storage::Storage::new(pool.clone());
+            let publisher = pub_sub.publisher();
+            let application_config = application_config.clone();
+            task::spawn(async move {
+                let node = SubxtNode::new(node_config)
+                    .await
+                    .context("create SubxtNode")?;
+                let sigterm =
+                    signal(SignalKind::terminate()).expect("SIGTERM handler can be registered");
+                chain_app::run(application_config.into(), node, storage, publisher, sigterm).await
+            })
+        };
+
+        let spo_indexer = {
+            let storage = spo_storage::Storage::new(pool.clone());
+            task::spawn(async move {
+                let node = SPOClient::new(spo_node_config.into())
+                    .await
+                    .context("create SPOClient")?;
+                let sigterm =
+                    signal(SignalKind::terminate()).expect("SIGTERM handler can be registered");
+                spo_app::run(spo_config.into(), node, storage, sigterm).await
+            })
+        };
+
+        let indexer_api = task::spawn({
+            let subscriber = pub_sub.subscriber();
+            let storage = api_storage::Storage::new(cipher.clone(), pool.clone());
+            let api = AxumApi::new(api_config, storage, subscriber.clone());
+
+            api_app::run(application_config.clone().into(), api, subscriber)
+        });
+
+        let wallet_indexer = task::spawn({
+            let storage = wallet_storage::Storage::new(cipher, pool);
+            let publisher = pub_sub.publisher();
+            let subscriber = pub_sub.subscriber();
+            let sigterm =
+                signal(SignalKind::terminate()).expect("SIGTERM handler can be registered");
+
+            wallet_app::run(
+                application_config.into(),
+                storage,
+                publisher,
+                subscriber,
+                sigterm,
+            )
+        });
+
+        let result = select! {
+            result = chain_indexer => task_result("chain-indexer", result),
+            result = spo_indexer => task_result("spo-indexer", result),
+            result = wallet_indexer => task_result("wallet-indexer", result),
+            result = indexer_api => task_result("indexer-api", result),
+        };
+
+        info!("indexer shutting down");
+        result
+    })
+}
+
+#[cfg(feature = "standalone")]
+fn task_result(
+    task_name: &str,
+    result: Result<anyhow::Result<()>, tokio::task::JoinError>,
+) -> anyhow::Result<()> {
+    use anyhow::anyhow;
+
+    match result {
+        Ok(Ok(())) => Err(anyhow!("{task_name} terminated")),
+        Ok(Err(error)) => Err(error.context(format!("{task_name} exited"))),
+        Err(error) => Err(anyhow!(error).context(format!("{task_name} panicked"))),
+    }
+}
